Fix warning.

+++ arch/mips/sgi-ip22/ip22-setup.c
@@ -140,7 +140,8 @@ void __init ip22_setup(void)
 #ifdef CONFIG_SGI_NEWPORT_CONSOLE
 	if (ctype && *ctype == 'g'){
 		unsigned long *gfxinfo;
-		long (*__vec)(void) = (void *) *(long *)((PROMBLOCK)->pvector + 0x20);
+		long (*__vec)(void) =
+			(void *) *(long *)(long)((PROMBLOCK)->pvector + 0x20);
 
 		gfxinfo = (unsigned long *)__vec();
 		sgi_gfxaddr = ((gfxinfo[1] >= 0xa0000000
